Exception handling

+++ src/main/java/com/curso/dscommerce/services/ProductService.java
@@ -1,16 +1,20 @@
 package com.curso.dscommerce.services;
 
-import java.util.List;
-
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.curso.dscommerce.dto.ProductDTO;
 import com.curso.dscommerce.entities.Product;
 import com.curso.dscommerce.repositories.ProductRepository;
+import com.curso.dscommerce.services.exceptions.DatabaseException;
+import com.curso.dscommerce.services.exceptions.ResourceNotFoundException;
+
+import jakarta.persistence.EntityNotFoundException;
 
 @Service
 public class ProductService {
@@ -20,7 +24,8 @@ public class ProductService {
 	
 	@Transactional(readOnly = true)
 	public ProductDTO findById(Long id) {
-		Product product = repository.findById(id).get();
+		Product product = repository.findById(id).orElseThrow(
+				() -> new ResourceNotFoundException("Recurso não encontrado"));
 		return new ProductDTO(product);
 	}
 	
@@ -42,15 +47,28 @@ public class ProductService {
 	
 	@Transactional
 	public ProductDTO update(Long id, ProductDTO dto) {
-		Product product = repository.getReferenceById(id);
-		copyDtoToProduct(dto, product);
-		product = repository.save(product);
-		return new ProductDTO(product);
+		try {
+			Product product = repository.getReferenceById(id);
+			copyDtoToProduct(dto, product);
+			product = repository.save(product);
+			return new ProductDTO(product);
+		} catch (EntityNotFoundException e) {
+			throw new ResourceNotFoundException("Recurso não encontrado");
+		}
 	}
 	
-	@Transactional
+	@Transactional(propagation = Propagation.SUPPORTS)
 	public void delete(Long id) {
-		repository.deleteById(id);
+		
+		if(!repository.existsById(id)) {
+			throw new ResourceNotFoundException("Recurso não encontrado");
+		}
+		
+		try {
+			repository.deleteById(id);
+		} catch (DataIntegrityViolationException e) {
+			throw new DatabaseException("Falha de integridade referencial");
+		}
 	}
 
 	private void copyDtoToProduct(ProductDTO dto, Product product) {
